Add: functionality to move items between slots
You can now move items between slots in the inventory.
This is done by clicking on a spell and then clicking on any other slot.
The item will be moved from the first slot to the second slot. If the second slot already contains an item, the items will be swapped.

This took way too long.